Fix for multi-queue scenarios with small batch sizes

+++ operators/advanced_network/advanced_network/managers/dpdk/adv_network_dpdk_mgr.h
@@ -148,7 +148,7 @@ class DpdkMgr : public Manager {
   void run() override;
   static constexpr int JUMBOFRAME_SIZE = 9100;
   static constexpr int DEFAULT_NUM_TX_BURST = 256;
-  static constexpr int DEFAULT_NUM_RX_BURST = 64;
+  static constexpr uint16_t DEFAULT_NUM_RX_BURST = 64;
   uint16_t default_num_rx_desc = 8192;
   uint16_t default_num_tx_desc = 8192;
   int num_ports = 0;
